bend_pvm::compiler::polkavm

Module host

Enums

HostFunction

This module defines the host functions that the PolkaVM runtime provides to contract executables. These functions allow the contract to interact with the blockchain environment.

Functions

generate_abi_helpers

Generate a helper function for ABI encoding function calls

generate_host_bindings

Generates bindings for host functions

generate_prelude

Generate prelude for all contracts