bend_pvm::compiler::polkavm::host

Enum HostFunction

#[repr(u32)]
pub enum HostFunction {

    StorageGet = 0,
    StorageSet = 1,
    StorageClear = 2,
    GetCaller = 10,
    GetCallValue = 11,
    GetBlockNumber = 12,
    GetBlockTimestamp = 13,
    Call = 20,
    StaticCall = 21,
    DelegateCall = 22,
    Create = 23,
    Create2 = 24,
    Keccak256 = 30,
    Blake2b256 = 31,
    Sha256 = 32,
    Ripemd160 = 33,
    EcdsaRecover = 34,
    Log = 40,
    Debug = 41,
    MemoryAlloc = 50,
    MemoryFree = 51,
    Abort = 60,
    Return = 61,
    Revert = 62,

}

This module defines the host functions that the PolkaVM runtime provides to contract executables. These functions allow the contract to interact with the blockchain environment.

Variants

StorageGet = 0

StorageSet = 1

StorageClear = 2

GetCaller = 10

GetCallValue = 11

GetBlockNumber = 12

GetBlockTimestamp = 13

Call = 20

StaticCall = 21

DelegateCall = 22

Create = 23

Create2 = 24

Keccak256 = 30

Blake2b256 = 31

Sha256 = 32

Ripemd160 = 33

EcdsaRecover = 34

Log = 40

Debug = 41

MemoryAlloc = 50

MemoryFree = 51

Abort = 60

Return = 61

Revert = 62